Patch 1/3 from David:
un/picklers: simplify the flattened object format

pickler.py previously contained this comment regarding flatten():
"""
this will place items in a sub dictionary (arguably not a pure JSON
representation, since it should be at root level.  However, this
method preserves the object so that it can be recreated as a
Python object
"""

We have done away with the need for the classdictitems__
sub-dictionary, thus simplifying the flattened format.

This was done by treating all values the same and relying on the
recursive correctness of flatten().