Add support for pickling recursive structures
Self-referential structures would previously fail to encode due to
infinite recursion caused by the cyclical structure of the objects
being encoded.

It is now possible to encode cyclical structures.
We accomplish this by maintaining independent object stacks in both the
pickler and unpickler.  The recursive flatten() method now detects cycles
and places object reference markers in the json data instead of
recursing into an object that has already been visited.

The recursive restore() tracks objects as they are created and resolves
the appropriate reference tags when they are detected in the json data.

jsonpickle_test.py was updated with tests that ensure that cyclical
objects are properly encode and decoded.

Signed-off-by: David Aguilar <davvid@gmail.com>