pickler: handle pickling dictionaries with non-string keys
It is possible that a dictionary with non-string keys are passed
to jsonpickle.  For example, it's legal for a dictionary to have
tuples as keys in Python.

JSON does not allow non-unicode keys, so for now we convert non-str
and non-unicode keys to unicode before marshalling off to the JSON
backend.

A better long term solution will be to implement dictionaries with
complex keys.  That will require some delicate jsonpickle surgery
which is why we saved that feature for later.

Signed-off-by: David Aguilar <davvid@gmail.com>