jsonpickle: Support objects with __setstate__() and __getstate__()
This adds a __getstate__() / __setstate__() extension mechanism
ala the Python stdlib cPickle module.  This allows objects with
properties or other C extensions to provide their own mechanism
for generating a json-friendly dictionary.

Instead of serializing the object directly, we now call the
object's __getstate__() method when flattening the object.
When restoring the object, we call its corresponding
__setstate__() method.

Signed-off-by: Dan Buch <daniel.buch@gmail.com>
Signed-off-by: David Aguilar <davvid@gmail.com>